notworkingone/backend/requirements.txt
flask
flask-cors
passlib[bcrypt]
pyjwt
requests
bcrypt==4.0.1
